requirements.txt
litellm
typer
rich
datasets
psutil
faiss-cpu
chromadb
qdrant-client
pymilvus
lancedb
pyarrow
